+++ src/Mediary/Core/Extensions/RequestExtensions.cs
@@ -28,26 +28,5 @@ public static class RequestExtensions
     public static RequestInfoAttribute? GetInfo<TResponse>(
         this IRequest<TResponse> request
     ) => RequestInfo.Get(request.GetType());
-
-    /// <summary>
-    /// Gets the description of a non-generic request from the <see cref="RequestInfoAttribute"/>, if available.
-    /// </summary>
-    public static string? GetDescription(
-        this IRequest request
-    ) => RequestInfo.GetDescription(request.GetType());
-
-    /// <summary>
-    /// Gets the tags of a non-generic request from the <see cref="RequestInfoAttribute"/>, if available.
-    /// </summary>
-    public static string[] GetTags(
-        this IRequest request
-    ) => RequestInfo.GetTags(request.GetType());
-
-    /// <summary>
-    /// Gets the <see cref="RequestInfoAttribute"/> instance from a non-generic request, if applied.
-    /// </summary>
-    public static RequestInfoAttribute? GetInfo(
-        this IRequest request
-    ) => RequestInfo.Get(request.GetType());
 }