+++ src/Mediary/Core/ICommand.cs
@@ -1,12 +1,6 @@
 ﻿namespace Mediary.Core;
 
 /// <summary>
-/// Represents a write command that does not return a value.
-/// Typically used to change system state (create, update, delete).
-/// </summary>
-public interface ICommand : IRequest { }
-
-/// <summary>
 /// Represents a write command that returns a response of type <typeparamref name="TResponse"/>.
 /// Typically used to change system state and return a result.
 /// </summary>